services/api/requirements.txt
# services/api/requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
simpleeval==0.9.13  # Safe math evaluation

# Async Database & Cache
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1

# AI & LLM Clients
openai==1.10.0   # Standard client often used for compatible endpoints
anthropic==0.8.0 # If using Claude fallback
tiktoken==0.5.2
sentence-transformers==2.3.1
transformers==4.37.0 # For tokenizer templates

# Graph & Vector DBs
neo4j==5.16.0
qdrant-client==1.7.3

# Agent Framework
langchain==0.1.5
langgraph==0.0.21

# Observability & Ops
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0
python-json-logger==2.0.7
backoff==2.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
boto3==1.34.34
httpx==0.26.0
tenacity==8.2.3